439 - Add mixins folder, a test mixin (wayne_commission), and a test spider (wayne_audit)
