+++ city_scrapers/spiders/wayne_audit.py
@@ -1,23 +1,19 @@
 # -*- coding: utf-8 -*-
-import re
-from datetime import datetime
-from dateutil.parser import parse as dateparse
-from urllib.parse import urljoin
 
+# WE ARE BRINGING IN A MIXIN WHICH IMPORTS OTHER LIBRARIES.
+# MIXINS ARE STORED IN /city-scrapers/city-scrapers/mixins
+# YOU CAN TAKE THE DEFINITIONS OUT OF THE MIXIN AND ADD THEM HERE IF THEY ARE
+# UNIQUE.
 
 from city_scrapers.spider import Spider
+from city_scrapers.mixins.wayne_commission import Wayne_commission
 
 
-class Wayne_auditSpider(Spider):
+class Wayne_auditSpider(Wayne_commission, Spider):
     name = 'wayne_audit'
     long_name = 'Wayne County Audit Committee'
     agency_id = 'Wayne County Audit Committee'
-    timezone = 'America/Detroit'
-    allowed_domains = ['www.waynecounty.com']
     start_urls = ['https://www.waynecounty.com/elected/commission/audit.aspx']
-
-    # Calendar shows only meetings in current year.
-    yearStr = datetime.now().year
 
     def parse(self, response):
         """
@@ -49,19 +45,6 @@ class Wayne_auditSpider(Spider):
             yield data
 
     @staticmethod
-    def _parse_documents(item, base_url):
-        url = item.xpath('td/a/@href').extract_first()
-        url = urljoin(base_url, url) if url is not None else ''
-        if url != '':
-            note = item.xpath('td/a/text()').extract_first()
-            note = note.lower() if note is not None else ''
-            return [{
-                'url': url,
-                'note': note
-            }]
-        return []
-
-    @staticmethod
     def _parse_description(response):
         """
         Event description taken from static text at top of page.
@@ -69,33 +52,6 @@ class Wayne_auditSpider(Spider):
         desc_xpath = '//h2[contains(text(), "Audit")]/following-sibling::div/section/p/text()'
         desc = response.xpath(desc_xpath).extract_first()
         return desc
-
-    def _parse_start(self, item):
-        """
-        Parse start date and time.
-        """
-        # Dateparse can't always handle the inconsistent dates, so
-        # let's normalize them using scrapy's regular expressions.
-        month_str = item.xpath('.//td[2]/text()').re(r'[a-zA-Z]{3}')[0]
-        day_str = item.xpath('.//td[2]/text()').re(r'\d+')[0]
-        time_str = item.xpath('.//td[3]/text()').extract_first()
-        date_str = dateparse('{0} {1} {2} {3}'.format(month_str, day_str, self.yearStr, time_str))
-
-        return {'date': date_str.date(), 'time': date_str.time(), 'note': ''}
-
-    def _parse_status(self, item, data):
-        """
-        Parse or generate status of meeting.
-        Postponed meetings will be considered cancelled.
-        """
-
-        status_str = item.xpath('.//td[4]/a/text() | .//td[4]/text()').extract_first()
-        # If the agenda column text contains "postponed," we consider it cancelled.
-        if re.search(r'postpone', status_str, re.IGNORECASE):
-            return 'cancelled'
-        # If it's not cancelled, use the status logic from spider.py
-        else:
-            return self._generate_status(data, '')
 
     @staticmethod
     def _parse_location():
